HW/HW2/requirements.txt
pandas==0.23.4
numpy==1.16.3
matplotlib==2.1.0
kneed==0.3.1
opencv-python==4.0.0.21
umap-learn==0.3.7
scikit-learn==0.20.3
tensorflow==1.13.1
glob3

